Address comments

+++ ambry-store/src/main/java/com/github/ambry/store/BlobStoreCompactor.java
@@ -43,6 +43,7 @@ import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
@@ -103,8 +104,8 @@ class BlobStoreCompactor {
   private volatile CountDownLatch runningLatch = new CountDownLatch(0);
   private byte[] bundleReadBuffer;
   private final AtomicReference<CompactionDetails> currentCompactionDetails = new AtomicReference();
-  private final AtomicLong compactedLogCount = new AtomicLong(0L);
-  private final AtomicLong logSegmentCount = new AtomicLong(0L);
+  private final AtomicInteger compactedLogCount = new AtomicInteger(0);
+  private final AtomicInteger logSegmentCount = new AtomicInteger(0);
 
   /**
    * Constructs the compactor component.
@@ -1160,7 +1161,7 @@ class BlobStoreCompactor {
           // eg: log segment under compaction [0_23, 130_16, 144_3]
           //     log segment after compaction [0_24, 130_24, 155_0, 166_0]
           // log segment 144_3's position doesn't exist in the compaction log, so we have one log segment compacted.
-          compactedLogCount.set(logSegmentPositionsUnderCompaction.stream()
+          compactedLogCount.set((int)logSegmentPositionsUnderCompaction.stream()
               .filter(p -> !logSegmentPositionsAfterCompaction.contains(p))
               .count());
         }
